+++ plugin.rb
@@ -14,6 +14,7 @@ require_relative 'engine/musicbrainz_place_onebox'
 require_relative 'engine/musicbrainz_recording_onebox'
 require_relative 'engine/musicbrainz_release_onebox'
 require_relative 'engine/musicbrainz_releasegroup_onebox'
+require_relative 'engine/musicbrainz_series_onebox'
 require_relative 'engine/musicbrainz_work_onebox'
 
 enabled_site_setting :musicbrainz_load_wikimedia_images