+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 0.8.0 - 202?-??-??
-- Show link to release group on CirtiqueBrainz in release onebox
+- Show link to release group on CritiqueBrainz in release onebox
 - DiscID onebox
+- Series onebox
 
 # 0.7.0 - 2021-10-18
 - Fixed display of label logos